reorganize handlers and tests
